refactor cloner-startup
