+++ images/cdi-cloner/cloner-startup/internal/helpers/env_process.go
@@ -1,0 +1,39 @@
+/*
+Copyright 2025 Flant JSC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package helpers
+
+import (
+	"fmt"
+	"os"
+	"strconv"
+)
+
+func GetEnv(key string) (string, error) {
+	value, ok := os.LookupEnv(key)
+	if !ok {
+		return "", fmt.Errorf("missing required environment variable: %s", key)
+	}
+	return value, nil
+}
+
+func GetBoolEnv(key string) (bool, error) {
+	value, err := GetEnv(key)
+	if err != nil {
+		return false, err
+	}
+	return strconv.ParseBool(value)
+}
